software/python-server/requirements.txt
Flask==3.0.3
google-generativeai==0.7.2
faster-whisper==1.0.3
huggingface-hub==0.25.1
requests==2.32.3
